move falls scripts and data, update barrier calculations, maybe fix #165

+++ scripts/prep/barriers/falls/falls.sh
@@ -1,25 +1,44 @@
 #!/bin/bash
-set -euxo pipefail
+
+# ----------
+# Script to:
+# - load falls from various sources
+# - match pts to streams
+# - combine falls into bcfishpass.falls
+# - identify barriers / non-barriers
+# ----------
+
+set -exo pipefail
+
+# check that path to data folder is provided
+if [ -z "$1" ]
+  then
+    echo "Usage: falls.sh </path/to/data/files>"
+    exit 1
+fi
 
 tmp="${TEMP:-/tmp}"
 
-# Script to:
-# - load falls from FISS and falls.geojson,
-# - match pts to streams
-# - combine falls into bcfishpass.falls
-
-# Archive existing FISS obstacles data just in case we want it later, the unique IDs aren't stable
-# (only keep one archive per day, and some work could be done to ensure the archive is only done if data has changed)
-# TODO - this presumes that the obstacles table is already present... will break on first run of script
-DOWNLOAD_DATE=$(psql -tc "SELECT replace(DATE(date_downloaded)::text, '-', '') FROM public.bcdata WHERE table_name = 'whse_fish.fiss_obstacles_pnt_sp'" | sed -e 's/^[[:space:]]*//')
-psql -c "DROP TABLE IF EXISTS whse_fish.fiss_obstacles_pnt_sp_$DOWNLOAD_DATE"
-psql -c "CREATE TABLE whse_fish.fiss_obstacles_pnt_sp_$DOWNLOAD_DATE AS SELECT * FROM whse_fish.fiss_obstacles_pnt_sp"
+# Archive existing FISS obstacles data just in case we want it later,
+# the unique IDs aren't stable. Note that the archive is a full copy and
+# will be made every time this is run on a different day
+if psql -c "SELECT to_regclass('whse_fish.fiss_obstacles_pnt_sp')" | grep -q 'whse_fish.fiss_obstacles_pnt_sp'; then
+  DOWNLOAD_DATE=$(psql -tc \
+    "SELECT replace(DATE(date_downloaded)::text, '-', '') \
+    FROM public.bcdata \
+    WHERE table_name = 'whse_fish.fiss_obstacles_pnt_sp'" \
+    | sed -e 's/^[[:space:]]*//')
+  # just drop existing archive for given date if it aleady exists and re-archive
+  psql -c "DROP TABLE IF EXISTS whse_fish.fiss_obstacles_pnt_sp_$DOWNLOAD_DATE"
+  psql -c "CREATE TABLE whse_fish.fiss_obstacles_pnt_sp_$DOWNLOAD_DATE AS SELECT * FROM whse_fish.fiss_obstacles_pnt_sp"
+fi
 
 # load the latest obstacles data
 bcdata bc2pg WHSE_FISH.FISS_OBSTACLES_PNT_SP
 
 # load additional (unpublished) obstacle data (provided by the Province, 2014)
-wget --trust-server-names -qNP "$tmp" https://hillcrestgeo.ca/outgoing/public/whse_fish/whse_fish.fiss_obstacles_unpublished.csv.zip
+wget --trust-server-names -qNP "$tmp" \
+  https://hillcrestgeo.ca/outgoing/public/whse_fish/whse_fish.fiss_obstacles_unpublished.csv.zip
 unzip -qjun -d "$tmp" "$tmp/whse_fish.fiss_obstacles_unpublished.csv.zip"
 
 psql -c "DROP TABLE IF EXISTS bcfishpass.fiss_obstacles_unpublished;"
@@ -47,7 +66,7 @@ psql -c "CREATE TABLE bcfishpass.falls_barrier_ind
  watershed_group_code        text,
  reviewer                    text,
  notes                       text)"
-psql -c "\copy bcfishpass.falls_barrier_ind FROM 'data/falls_barrier_ind.csv' delimiter ',' csv header"
+psql -c "\copy bcfishpass.falls_barrier_ind FROM '$1/falls/falls_barrier_ind.csv' delimiter ',' csv header"
 
 
 # load other falls, from various sources (add any new falls to this table)
@@ -64,7 +83,7 @@ psql -c "CREATE TABLE bcfishpass.falls_other
    notes text,
    primary key (blue_line_key, downstream_route_measure)
    )"
-psql -c "\copy bcfishpass.falls_other FROM 'data/falls_other.csv' delimiter ',' csv header"
+psql -c "\copy bcfishpass.falls_other FROM '$1/falls/falls_other.csv' delimiter ',' csv header"
 
 # match falls to streams, combine sources
 psql -f sql/falls.sql
